// src/app/shared/model/CustPersonalFinDataObj.Model.ts
export class CustPersonalFinDataObj {
    CustPersonalFinDataId: any;
    CustPersonalId: any;
    MonthlyIncomeAmt: any;
    MonthlyExpenseAmt: any;
    MonthlyInstallmentAmt: any;
    MrSourceOfIncomeCode: any;
    SpouseMonthlyIncomeAmt: any;
    IsJoinIncome: any;
    RowVersion: any;
    constructor(){this.CustPersonalFinDataId = 0, this.RowVersion = ""}
}